Replace incomplete parameters file with script.
One of the difficulties for beginners with the template is the
copy-paste necessary for adding the security certificate ID to the
CloudFormation parameter file. This is particularly unfortunate, as the
action is both mindless, cryptic, and error prone.

Instead, this commit removes the need for this action, introducing a
script that generates the parameter file once the security certificate
is uploaded to AWS.